+++ lua/autorun/ttt_deadringer.lua
@@ -47,12 +47,14 @@ end
 
 CreateConVar('ttt_deadringer_chargetime', '20', flags, 'Time it takes to recharge the Dead Ringer.')
 CreateConVar('ttt_deadringer_cloaktime', '7', flags, 'Time that the Dead Ringer will cloak you for.')
+CreateConVar('ttt_deadringer_cloaktime_reuse', '0', flags, 'Whether or not the Dead Ringer will convert unused cloak time into charge time.')
+
 CreateConVar('ttt_deadringer_damage_reduction', '0.65', flags, 'Damage reduction while cloaked.')
 CreateConVar('ttt_deadringer_damage_reduction_time', '3', flags, 'Damage reduction time while cloaked. (1.0 = 1 second of damage reduction)')
 CreateConVar('ttt_deadringer_damage_reduction_initial', '0.75', flags, 'Percentage of damage reduction for the initial hit which triggers the Dead Ringer. (0.75 = 75%)')
-CreateConVar('ttt_deadringer_cloaktime_reuse', '0', flags, 'Whether or not the Dead Ringer will convert unused cloak time into charge time.')
-CreateConVar('ttt_deadringer_corpse_role', '0', flags, 'Whether or not the Dead Ringer will show the real role of the player on the corpse.')
-CreateConVar('ttt_deadringer_corpse_confirm', '1', flags, 'Whether or not the Dead Ringer will confirm the death of the player on the corpse.')
+
+CreateConVar('ttt_deadringer_corpse_mode', '0', flags, 'Whether or not the Dead Ringer will show the real role of the player on the corpse. (0 = fake, 1 = real, 2 = innocent)')
+CreateConVar('ttt_deadringer_corpse_confirm', '0', flags, 'Whether or not the Dead Ringer will confirm the death of the player on the corpse. (Not recommended for TTT2)')
 
 -- Sandbox Compatibility
 
@@ -66,7 +68,9 @@ local DEADRINGER_DEFAULTS = {
 	ttt_deadringer_damage_reduction = 0.65,
 	ttt_deadringer_damage_reduction_time = 3,
 	ttt_deadringer_damage_reduction_initial = 0.75,
-	ttt_deadringer_cloaktime_reuse = 0
+	ttt_deadringer_cloaktime_reuse = 0,
+	ttt_deadringer_corpse_mode = 0,
+	ttt_deadringer_corpse_confirm = 0
 }
 
 hook.Add('PopulateToolMenu', 'DeadringerPopulateToolMenu', function()
@@ -101,11 +105,11 @@ hook.Add('PopulateToolMenu', 'DeadringerPopulateToolMenu', function()
 
 		panel:Help('Corpse Settings')
 
-		panel:CheckBox('Show Role', 'ttt_deadringer_corpse_role')
-		panel:ControlHelp('Whether or not the Dead Ringer will show the real role of the player on the corpse.')
+		panel:NumSlider('Role Mode', 'ttt_deadringer_corpse_mode', 0, 2, 0)
+		panel:ControlHelp('Whether or not the Dead Ringer will show the real role of the player on the corpse. (0 = fake, 1 = real, 2 = innocent)')
 
 		panel:CheckBox('Confirm Death', 'ttt_deadringer_corpse_confirm')
-		panel:ControlHelp('Whether or not the Dead Ringer will confirm the death of the player on the corpse.')
+		panel:ControlHelp('Whether or not the Dead Ringer will confirm the death of the player on the corpse. (Not recommended for TTT2)')
 	end)
 end)
 
@@ -138,8 +142,8 @@ if CLIENT then
 		LANG.AddToLanguage(lang, 'label_ttt_deadringer_cloaktime_reuse', 'Cloak Time Reuse')
 		LANG.AddToLanguage(lang, 'help_ttt_deadringer_cloaktime_reuse', 'Whether or not the Dead Ringer will convert unused cloak time into charge time.')
 
-		LANG.AddToLanguage(lang, 'label_ttt_deadringer_corpse_role', 'Show Role')
-		LANG.AddToLanguage(lang, 'help_ttt_deadringer_corpse_role', 'Whether or not the Dead Ringer will show the real role of the player on the corpse.')
+		LANG.AddToLanguage(lang, 'label_ttt_deadringer_corpse_mode', 'Role Mode')
+		LANG.AddToLanguage(lang, 'help_ttt_deadringer_corpse_mode', 'Whether or not the Dead Ringer will show the real role of the player on the corpse. (0 = fake, 1 = real, 2 = innocent)')
 
 		LANG.AddToLanguage(lang, 'label_ttt_deadringer_corpse_confirm', 'Confirm Death')
 		LANG.AddToLanguage(lang, 'help_ttt_deadringer_corpse_confirm', 'Whether or not the Dead Ringer will confirm the death of the player on the corpse.')
@@ -239,7 +243,7 @@ hook.Add('Initialize', 'DeadringerInitialize', function()
 			corpseList:SetSize(390, 150)
 			corpseList:SetSpacing(5)
 
-			local corpseRole = xlib.makecheckbox{label = 'Show Role (def. 0)', repconvar = 'rep_ttt_deadringer_corpse_role', parent = corpseList}
+			local corpseRole = xlib.makeslider{label = 'Role Mode (def. 0)', repconvar = 'rep_ttt_deadringer_corpse_mode', min = 0, max = 2, decimal = 0, parent = corpseList}
 			corpseList:AddItem(corpseRole)
 
 			local corpseConfirm = xlib.makecheckbox{label = 'Confirm Death (def. 1)', repconvar = 'rep_ttt_deadringer_corpse_confirm', parent = corpseList}
@@ -258,7 +262,7 @@ hook.Add('Initialize', 'DeadringerInitialize', function()
 			ULib.replicatedWritableCvar('ttt_deadringer_damage_reduction_time', 'rep_ttt_deadringer_damage_reduction_time', GetConVar('ttt_deadringer_damage_reduction_time'):GetFloat(), true, false, name)
 			ULib.replicatedWritableCvar('ttt_deadringer_damage_reduction_initial', 'rep_ttt_deadringer_damage_reduction_initial', GetConVar('ttt_deadringer_damage_reduction_initial'):GetFloat(), true, false, name)
 			ULib.replicatedWritableCvar('ttt_deadringer_cloaktime_reuse', 'rep_ttt_deadringer_cloaktime_reuse', GetConVar('ttt_deadringer_cloaktime_reuse'):GetBool(), true, false, name)
-			ULib.replicatedWritableCvar('ttt_deadringer_corpse_role', 'rep_ttt_deadringer_corpse_role', GetConVar('ttt_deadringer_corpse_role'):GetBool(), true, false, name)
+			ULib.replicatedWritableCvar('ttt_deadringer_corpse_mode', 'rep_ttt_deadringer_corpse_mode', GetConVar('ttt_deadringer_corpse_mode'):GetInt(), true, false, name)
 			ULib.replicatedWritableCvar('ttt_deadringer_corpse_confirm', 'rep_ttt_deadringer_corpse_confirm', GetConVar('ttt_deadringer_corpse_confirm'):GetBool(), true, false, name)
 		end)
 	end