+++ lua/terrortown/lang/en/fake.lua
@@ -1,0 +1,7 @@
+local L = LANG.GetLanguageTableReference('en')
+
+L[FAKE.name] = 'Fake'
+L[FAKE.defaultTeam] = 'Fake Team'
+L['body_found_' .. FAKE.abbr] = 'Their body was fake!'
+L['search_role_' .. FAKE.abbr] = 'This person is fake!'
+L['ttt2_desc_' .. FAKE.name] = [[A fake temporary role for players trying to fake their role to others (such as corpses, disguises, etc.)]]